hydrolix-cdn-insights/cdk-hydrolix-data-assistant-agentcore-strands/hydrolix-data-assistant-agentcore-strands/requirements.txt
strands-agents
strands-agents-tools
typing
boto3
bedrock-agentcore
bedrock-agentcore-starter-toolkit
botocore
mcp
clickhouse-connect
fastmcp
gunicorn